requirement.txt
Django==2.2
django-bootstrap4==0.0.8
django-braces==1.13.0
django-cors-headers==2.5.3
djangorestframework==3.9.2
multipledispatch==0.6.0
mysqlclient==2.1.1
numpy==1.21.6
Pillow==9.5.0
pygame==2.4.0
PyOpenGL==3.1.7
pyrr==0.10.3
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
turtle==0.0.1
typing_extensions==4.5.0
